requirements.txt
ariadne==0.26.2
Flask==3.1.1
flask_cors==6.0.1
Flask_JWT_Extended==4.7.1
flask_pymongo==3.0.1
marshmallow==4.0.0
pymongo==4.13.2
python-dotenv==1.1.1
Werkzeug==3.1.3
